// src/api/models/streamContentBlockDeltaDTO.ts
/**
 * Generated by orval v8.4.2 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { StreamDeltaPayloadDTO } from './streamDeltaPayloadDTO';

export interface StreamContentBlockDeltaDTO {
  type: string;
  delta: StreamDeltaPayloadDTO;
}