// src/api/models/streamMessageCompleteDTO.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Folioo API docs
 * API document of Folioo Development
 * OpenAPI spec version: 1.0.0
 */
import type { StreamMessagePayloadDTO } from './streamMessagePayloadDTO';

export interface StreamMessageCompleteDTO {
  type: string;
  message: StreamMessagePayloadDTO;
}